+++ src/Genesis/Hub.php
@@ -10,46 +10,27 @@ declare(strict_types=1);
 namespace DecodeLabs\Genesis;
 
 use DecodeLabs\Fluidity\Cast;
+use DecodeLabs\Genesis;
 use DecodeLabs\Genesis\Build\Manifest as BuildManifest;
 use DecodeLabs\Genesis\Environment\Config as EnvConfig;
-use DecodeLabs\Genesis\Loader\Stack as StackLoader;
+use DecodeLabs\Kingdom;
 
 interface Hub extends Cast
 {
     public ?BuildManifest $buildManifest { get; }
 
-    /**
-     * Init with options
-     */
     public function __construct(
-        Context $context,
+        Genesis $service,
         Bootstrap $bootstrap
     );
 
-    /**
-     * Register custom loaders into the stack
-     */
-    public function initializeLoaders(
-        StackLoader $loader
-    ): void;
+    public function initializeLoaders(): void;
 
-    /**
-     * Load build info
-     */
     public function loadBuild(): Build;
 
-    /**
-     * Load object to convey environment configuration
-     */
     public function loadEnvironmentConfig(): EnvConfig;
 
-    /**
-     * Setup platform libraries
-     */
     public function initializePlatform(): void;
 
-    /**
-     * Load kernel to handle running the app
-     */
-    public function loadKernel(): Kernel;
+    public function loadKingdom(): Kingdom;
 }